requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.16
django-bootstrap3==11.0.0
django-model-utils==3.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
nose==1.3.7
psycopg2==2.7.6.1
pudb==2018.1
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.7
static3==0.7.0
urwid==2.0.1
whitenoise==3.3.1
